Mock tensorboard callbacks for all freqAI tests
